RDKCOM-2759: REFPLTB-1510 : CcspMisc - Resolve all build warnings and errors
Reason for change: Have to clear warnings with respect to 64bit architecture
Test Procedure: build should be success on both 64bit and 32bit arch
Risks: Low

(cherry picked from commit 599691b95f5c5dc72f3764a6bbb5ab5f507deaf8)

+++ source/dhcp_client_utils/dhcp_client_common.c
@@ -313,7 +313,7 @@ static int parseArgs(const char *cmd, const char *args, char ***argv)
 
     if (array[argIndex] == NULL)
     {
-        DBG_PRINT("memory allocation of %d failed", strlen(cmdStr) + 1);
+        DBG_PRINT("memory allocation of %lu failed", (ULONG)strlen(cmdStr) + 1);
         freeArgs(array);
         return FAILURE;
     }
